chore: move AddressPriorityQueue to pool
Subsequent commits will use the switch the inner type of txs_by_account
and use it there, also rename.

commit-id:30696bc0
