+++ crates/mempool/src/transaction_queue.rs
@@ -1,8 +1,7 @@
 use std::cmp::Ordering;
-use std::collections::{BTreeSet, HashMap, VecDeque};
+use std::collections::{BTreeSet, HashMap};
 
 use starknet_api::core::{ContractAddress, Nonce};
-use starknet_mempool_types::mempool_types::ThinTransaction;
 
 use crate::mempool::TransactionReference;
 // Assumption: for the MVP only one transaction from the same contract class can be in the mempool
@@ -86,44 +85,5 @@ impl Ord for QueuedTransaction {
 impl PartialOrd for QueuedTransaction {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
-    }
-}
-
-// TODO: remove when is used.
-#[allow(dead_code)]
-// Invariant: Transactions have strictly increasing nonces, without gaps.
-// Assumption: Transactions are provided in the correct order.
-#[derive(Default)]
-pub struct AddressPriorityQueue(VecDeque<ThinTransaction>);
-
-// TODO: remove when is used.
-#[allow(dead_code)]
-impl AddressPriorityQueue {
-    pub fn push(&mut self, tx: ThinTransaction) {
-        if let Some(last_tx) = self.0.back() {
-            assert_eq!(
-                tx.nonce,
-                last_tx.nonce.try_increment().expect("Nonce overflow."),
-                "Nonces must be strictly increasing without gaps."
-            );
-        }
-
-        self.0.push_back(tx);
-    }
-
-    pub fn top(&self) -> Option<&ThinTransaction> {
-        self.0.front()
-    }
-
-    pub fn pop_front(&mut self) -> Option<ThinTransaction> {
-        self.0.pop_front()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    pub fn contains(&self, tx: &ThinTransaction) -> bool {
-        self.0.contains(tx)
     }
 }
